% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbsorptionGraphicFunction.R
\name{AbsorptionGraphic}
\alias{AbsorptionGraphic}
\title{Plotting absorption by plant and value stream}
\usage{
AbsorptionGraphic(x, labels, barwidth = 0.7, col = c(rgb(192, 0, 0,
  maxColorValue = 255), rgb(166, 166, 166, maxColorValue = 255)))
}
\arguments{
\item{x}{dataframe with 4 columns. The columns should be as follows: 
\itemize{
\item{Value Stream- these shoule be abbreviations.  The values will be used as labels on the chart}
\item{Comparable- this is likely the actuals or new forecast}
\item{Comparable- this is likely the forecast or old forecast}
\item{Plant- Abbreviation of the plant name.  One of Rar, Roc, or Penc} }}

\item{labels}{vector of lenghth two.  Labels for the stacked bars}

\item{barwidth}{width of bars}

\item{col}{color scheme}
}
\value{
ggpplot object
}
\description{
Creates a stacked bar chart of absorption numbers by value stream.  The plot is facetted by plant and provides 2 points for comparison.  The chart can be used to compare actuals against a forecast or two forecasts.
}
\examples{
dat<- data.frame(VS= paste0("VS", 1:3),
                 comp1= c(4:6)*1000,
                 comp2= c(5:7)*1000,
                 Plant= rep("Rar",3))
AbsorptionGraphic(dat, c("FCST","ACT"))
}
