% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FormatFITReport.R
\name{FormatFITReport}
\alias{FormatFITReport}
\title{Process raw FIT data for computation}
\usage{
FormatFITReport(file)
}
\arguments{
\item{file}{path to .csv file extracted from FIT.}
}
\description{
Preprocess an extract from the FIT financial reporting system and prepare
it for further analysis.  The basic premise is to clean the column headers
and remove any extraneous leading characters.
}
\details{
The data extract is created by using transaction code \code{
s_pl0_86000030}.  In the search paramenter input screen, use the following 
settings:
\itemize{
\item{Company Code: }{1001, 1002}
\item{Function: }{2001}
\item{Ledger: }{0L}
\item{Currency Type: }{10}
}
}
