% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ortho.R
\docType{package}
\name{Ortho}
\alias{Ortho}
\alias{Ortho-package}
\title{Ortho}
\description{
Providing common functions for working with financial data
at Ortho Clinical Diagnostics. The financial data is often 
exported from FIT- an Oracle general ledger system. The
extracts often do not have formatting conducive to analysis
in R so require additional munging. This package includes 
the convenience functions to perform this common task.
}
\author{
Ryan Kuhn
}
